Makes post process scripts able to pass args along with filename
Any whitespace is the boundrary between the args/filename. Whitespace can be
escaped by putting a exxclamation point in front of it. And exclamation points
can be escaped by putting an exclamation point in front of the exclamation point
to be escaped.

I thought about adding another box for arguments, but I think that would make it
more confusing to use. The only worry I have with this method is peoples existing
scripts with whitespace in the name.