Refer to manual in README
Once the CLI changes are updated and changed, I believe it is preferable to link to an authoritative source of information (the Manual), instead of creating a duplicate that needs to be maintained separately. 

This is an application of the DRY principle (Don't Repeat Yourself).

To be merged after https://github.com/slic3r/Slic3r-Manual/pull/60 is merged.