Change datadir to use GetUserConfigDir()
Since there are only stored configuration files it makes sense to use the appropriate folder.

This has the advantage over GetUserDataDir() that the XDG-specification can be used on Unix. Also the folder for Mac changes to an equivalent.

The following folders will be used with the applications name appended:

* Unix: XDG_CONFIG_HOME or ~/.config
* Windows: C:\Users\username\AppData\Roaming or C:\Documents and Settings\username\Application Data
* Mac: ~/Library/Preferences

If the folder exists at the old location it will be used instead.

These changes require wxWidgets >= 3.1.1