Cover coInt situation for configuration options. Wire in reloading configs on initial load. Implement deserialize method on UI_Fields.
PresetChooser: Use a view of the underlying map. Add LogChannel()

Add a method that gets a reference-wrapped view of Preset objects from the main storage.

misc_ui: Add a simple implementation of std::erase_if for std::map if not building on c++20

GUI: Adapt to use a map instead of a vector.
Preset: make preset_store an array of maps so we can keyword-reference.
PresetPage: add_optgroup
PresetEditor: add methods to find the currently used preset (using -default- if nothing).