Fix crash on FreeBSD
On FreeBSD builds, if you try to change a printer/filament/setting
option while the dirty flag is set, it crashes.  This patch seems
to fix the crash, though I'm not sure what the id is actually used
for.

merill: added the #ifdef __WXGTK__ to not mess with other OSes