Add support for libspnav
When using a 3DXonnexion mouse on Linux or FreeBSD systems, USB
HID mode is being used.  This means only one program can be running
at a time with Space Mouse support, and potentially means permissions
issues for USB devices.

This patch uses libspnav if available to avoid these issues.  This
also has the advantage of not relying on the device driver to be
updated in Slic3r/Prusa Slicer/Super Slicer whenever a new device
is supported.